Move WidgetFrame sub-components
