+++ static/app/views/dashboards/widgetCard/tooltipIconTrigger.tsx
